# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autores",
        "description": "API para gerenciar autores",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/authors/store": {
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Cria um novo autor",
                "operationId": "218706a81c6fdf1db119288acfca1d43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "birth_date"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do autor",
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "description": "1981-10-21",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Record saved successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID do autor",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Nome do autor",
                                                    "type": "string"
                                                },
                                                "birth_date": {
                                                    "description": "1981-10-21",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error trying to save."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{id}/show": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Exibe os detalhes de um autor",
                "operationId": "52824161f47da650d6eff99372d0f43e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do autor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do autor",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Nome do autor",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{id}/update": {
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Atualiza os dados de um autor",
                "operationId": "9aa357039f8f29fe2d4a1cb421731568",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do autor",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Record changed successfully!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID do autor",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Nome do autor",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error trying to update."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{id}/destroy": {
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Remove um autor",
                "operationId": "3ce0fe603e5f967ccdd4a4a8648b7ae7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor removido com sucesso"
                    },
                    "500": {
                        "description": "Erro ao remover",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error trying to delete."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}